TrialProjects/LLM-Test/requirements.txt
ollama
pyodbc
flask
flask-cors
python-dotenv